# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Claw-Empire API",
    "version": "1.2.4",
    "description": "Baseline OpenAPI contract for commonly used Claw-Empire endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin (recommended for local docs)"
    },
    {
      "url": "http://127.0.0.1:8790",
      "description": "Local development (loopback)"
    },
    {
      "url": "http://localhost:8790",
      "description": "Local development (localhost)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_AUTH_TOKEN"
      },
      "inboxSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-inbox-secret"
      }
    },
...
  },
  "paths": {
...
    "/api/tasks/{id}/run": {
      "post": {
        "tags": ["tasks"],
        "summary": "Run task",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task run started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "unauthorized",
                  "message": "Authentication is required."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "forbidden",
                  "message": "You do not have permission to access this resource."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "conflict",
                  "message": "The request conflicts with current server state."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "too_many_requests",
                  "message": "Rate limit exceeded."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "internal_error",
                  "message": "Unexpected server error."
                }
              }
            }
          }
        }
      }
    },
...
  }
}